+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/util/annotation/ApiOperation.java
@@ -1,0 +1,13 @@
+package com.pt.pedrovcruzeiro.phonebook.util.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ApiOperation {
+
+  String value();
+}