+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/util/error/InvalidPhoneNumberException.java
@@ -1,8 +1,0 @@
-package com.pt.pedrovcruzeiro.phonebook.util.error;
-
-public class InvalidPhoneNumberException extends RuntimeException {
-
-  public InvalidPhoneNumberException(String entity) {
-    super(entity);
-  }
-}