+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/util/error/RetrieveBusinessSectorException.java
@@ -1,0 +1,8 @@
+package com.pt.pedrovcruzeiro.phonebook.util.error;
+
+public class RetrieveBusinessSectorException extends RuntimeException {
+
+  public RetrieveBusinessSectorException(String entity) {
+    super(entity);
+  }
+}