+++ src/test/java/com/pt/pedrovcruzeiro/phonebook/service/BusinessSectorServiceTest.java
@@ -1,3 +1,3 @@
 package com.pt.pedrovcruzeiro.phonebook.service;
 
-class PrefixesServiceTest {}
+class BusinessSectorServiceTest {}