+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All technical changes ill be documented in this file.
 
 ## [Unreleased]
 
+## [1.0.0] - 2021-12-10
+
+### Added
+
+* Added Service unit tests
+* Added controller advice
+* Added MDC Context interceptor
+
+### Changed
+
+* Improved overall code quality with exception handling
+
+
 ## [0.0.2] - 2021-12-09
 
 ### Added